# lib/api-zod/src/generated/types/openaiConversation.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * AI Diet Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: Date;
}